aplicando los consejos de ger en el islogged middleware
