+++ server/controllers/auth.js
@@ -1,5 +1,6 @@
 import UserModel from "../models/User";
 
+
 const login = (req, res) => {
   if (!req.body.password || !req.body.email) {
     return res.status(400).send("Debe proporcionar usuario y contraseña");
@@ -15,6 +16,28 @@ const login = (req, res) => {
     });
 };
 
+
+// middlewares
+
+const isLogged = (req, res, next) => {
+
+	/* recibimos el _id de mongo como si fuera un token, 
+	tener en cuenta que mongo explota si le pasas un id incorrecto */
+
+	const token = req.get('Autorization');
+
+	UserModel.findOne({ _id: token })
+		.then((user) => {
+			user && next()
+			res.status(403).json('no tenes permisos');
+		})
+		.catch((error) => {
+			res.status(500).json(error);
+		})
+}
+
+
 export default {
   login,
+  isLogged,
 };
